% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binom.low}
\alias{binom.low}
\title{binom.low}
\usage{
binom.low(x, n)
}
\arguments{
\item{x}{number of 1's observed that year}

\item{n}{total observations in that year}
}
\description{
use these functions to make error bars with binomial test
}
\examples{
\dontrun{
 
}
}
