% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binom_plot}
\alias{binom_plot}
\title{binom_plot
use this to make plots of a binomial variable (one that takes only 0/1 values)}
\usage{
binom_plot(f, label = "", label_y = 0)
}
\arguments{
\item{f}{a data frame that must contain the following columns:
year = year of observations
x = number of 1's observed that year
n = total observations in that year}
}
\description{
binom_plot
use this to make plots of a binomial variable (one that takes only 0/1 values)
}
\examples{
\dontrun{
 
}
}
