% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-analysis.R
\name{build_analysis_package}
\alias{build_analysis_package}
\title{Build the analysis project}
\usage{
build_analysis_package(form = "form.xlsx")
}
\arguments{
\item{form}{The full file name of the form to be accessed (not that it has to be an .xls format).}
}
\description{
Function to build the analysis project.
In order to start, you need to put both the data and the original xlsform within the data-raw folder
you can adjust the function below and add as many output format as required
}
\examples{
\dontrun{
build_analysis_package(form = "form.xlsx")
}
}
