% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_data_bin}
\alias{cor_data_bin}
\title{cor_data_bin}
\usage{
cor_data_bin(d, my_var, bin_vars, ord_vars, idx_vars, unord_vars)
}
\arguments{
\item{f}{data frame f needs to contain a year column all other columns will
become the names of data series that will be plotted together}
}
\description{
cor_data_bin
}
\examples{
\dontrun{
 
}
}
