% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geo_variance}
\alias{geo_variance}
\title{geo_variance}
\usage{
geo_variance(data, x, make_plot = FALSE, xlab = "Value")
}
\arguments{
\item{data}{data}

\item{x}{must be the output from a multiple regression on MICE output}

\item{make_plot}{bollean}

\item{xlab}{plot title}
}
\description{
ggplot2 object with results of a regression
}
\examples{
\dontrun{

}
}
