% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataLAPOP.R
\name{get_dataLAPOP}
\alias{get_dataLAPOP}
\title{get_dataLAPOP}
\usage{
get_dataLAPOP()
}
\arguments{
\item{src}{list of direct URL to dataset - dataset name to be normalised
with 3 letters country code plus yer
like GTM2019}
}
\description{
Get all country specific dataset
}
\examples{
\dontrun{
}
}
