% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_outliers}
\alias{get_outliers}
\title{get_outliers}
\usage{
get_outliers(var, direction, pmax = 0.01)
}
\arguments{
\item{var}{variable to look at}

\item{direction}{should be 'less', 'greater', or 'two.sided'}

\item{pmax}{max value}
}
\description{
For the above-average areas, which individual cities are above average?
This often will not return anything.
}
\examples{
\dontrun{
get_outliers('fear_idx',g,pmax=0.05)
}
}
