% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lm_plot}
\alias{lm_plot}
\title{lm_plot}
\usage{
lm_plot(
  x,
  orig,
  title,
  p_max = 0.01,
  text = NULL,
  offsets = NULL,
  flip = NULL,
  width = 15
)
}
\arguments{
\item{x}{must be the output from a multiple regression on MICE output}

\item{orig}{must be the original, unimputed datasetas}

\item{p_max}{is the maximum p-value that will be included in the viz}

\item{text}{is a 2-column data frame containing variable names and
the text with which they should be replaced}

\item{offsets}{is a 3-column data frame (var,x,y) detailing how
labels should be offset for readability.}

\item{flip}{is a list of variables whose directionality should be}

\item{width}{is the width (in characters) to be used for wrapping of long labels}
}
\description{
ggplot2 object with results of a regression
}
\examples{
\dontrun{

}
}
