% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataLAPOP.R
\name{make_codebook}
\alias{make_codebook}
\title{make_codebook}
\usage{
make_codebook(tbl)
}
\arguments{
\item{tbl}{labeled dataframe}
}
\description{
Create a codebook
https://stackoverflow.com/questions/39671621/extract-the-labels-attribute-from-labeled-tibble-columns-from-a-haven-import-f
}
\examples{
\dontrun{
elsalvador2018 <- read_dta("http://datasets.americasbarometer.org/database/files/ElSalvador\%20LAPOP\%20AmericasBarometer\%202018\%20v1.0_W.dta")
elsalvador2018_codebook <- make_codebook(elsalvador2018)
}
}
