% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_idx}
\alias{make_idx}
\title{make_idx}
\usage{
make_idx(data, vars, sgn = 1, scale = FALSE, seed = 12345)
}
\arguments{
\item{data}{Given dataset containing variables}

\item{vars}{variable used construct a composite index from the first principal
component after multiple imputation.}

\item{sgn}{Direction -  Set sgn=-1 to reverse the direction of the index.}

\item{scale}{Scale - Set scale=TRUE to use scaling during the PCA calculation}

\item{seed}{for reproducibility}
}
\description{
Functions we need to construct indices
}
\examples{
\dontrun{
 
}
}
