% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataLAPOP.R
\name{merge_dataLAPOP}
\alias{merge_dataLAPOP}
\title{merge_dataLAPOP}
\usage{
merge_dataLAPOP()
}
\description{
Do a full merge and get the
full Grand Merge data set.
Create a data frame called lapop.trends containing all indicators
that are consistent across countries and years.
}
\examples{
\dontrun{
}
}
