% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanity_check}
\alias{sanity_check}
\title{sanity_check}
\usage{
sanity_check(data, idx, var)
}
\arguments{
\item{data}{Given dataset containing variables}

\item{idx}{Consutructed Index}

\item{var}{variable used construct a composite index from the first principal}
}
\description{
Return the average values of 'var' in 'data' for the lowest and highest quartiles of idx
}
\examples{
\dontrun{
 
}
}
