% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trend_plot}
\alias{trend_plot}
\title{trend_plot
use this to make plots of a binomial variable (one that takes only 0/1 values)}
\usage{
trend_plot(f, plotTittle = "", plotsubTittle = "")
}
\arguments{
\item{f}{a data frame that must contain the following columns:
year = year of observations
x = number of 1's observed that year
n = total observations in that year}

\item{plotTittle}{plot title}

\item{plotsubTittle}{plot title}
}
\description{
trend_plot
use this to make plots of a binomial variable (one that takes only 0/1 values)
}
\examples{
\dontrun{
 
}
}
