% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bar_plot}
\alias{bar_plot}
\title{bar_plot}
\usage{
bar_plot(f, plotTittle = "", plotsubTittle = "")
}
\arguments{
\item{f}{data frame f needs to contain a year column all other columns will
become the names of data series that will be plotted together}

\item{plotTittle}{plot title}

\item{plotsubTittle}{plot title}
}
\description{
ggplot2 bar plot: plot distribution of x (no time resolution here)
use this to make bar plots of a categorical indicator
with >2 categories
}
\examples{
\dontrun{
 
}
}
