% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataLAPOP.R
\name{get_dataLAPOP}
\alias{get_dataLAPOP}
\title{get_dataLAPOP}
\usage{
get_dataLAPOP()
}
\description{
Get all country specific dataset and do a full merge and get the The full Grand Merge data set
}
\examples{
\dontrun{
}
}
